chore: test to see if vercel allows deploy on src
