--- api/index.d.ts
@@ -1,5 +1,0 @@
-import express from 'express';
-
-declare const _default: express.Express;
-
-export { _default as default };
