+++ package.json
@@ -5,7 +5,7 @@
   "main": "api/index.js",
   "scripts": {
     "dev": "dotenv -e .env ts-node src/index.ts",
-    "build:api": "bunchee",
+    "build": "bunchee",
     "start": "dotenv -e .env node api/index.js"
   },
   "license": "MIT",
